refactor(control): split event/command types into control/events.rs (Phase 6)
control.rs becomes control/mod.rs; AgentEvent (the unified event
stream), TextOp, AgentTreeNode, ControlCommand, UserMessageAck, and
MissionStatus move verbatim to control/events.rs with a glob re-export
preserving every path. First mechanical cut of the control.rs split —
streaming/queue/actor/snapshot extractions follow the same seam. No
behavior change.
